Begin switch to bazel from MAKE.
This is to more easily enable cross-platform building and to keep me
sane since at work I am effectively using bazel, and haven't used MAKE
in ages.  While this does now require the supporting system to have
bazel installed, this is much preferable to having two completley
separate build configs per platform (linux and darwin).

In order to enable this, I've added some bazel specific files, as well
as temporarily disabled problem 7, which relies on an openssl integration
that is not yet supported.

Testing is also not yet supported.