Fix signed-ness bug in RandomBlob generation.
The original code relied on using 'len' as a counter maintaining the remaining
number of bytes to copy, but removed potentially more from the counter than the
counter contained.  This was thought to be ok, with the condition just checking
to see if the counter was positive, but since 'len' is of type size_t, it is
unsigned, and therefore would never be negative.  This could yield an infinite
loop any time the desired RandomBlob length was not a multiple of the size of
each chunk of random bytes obtained.

Fix this by subracting exactly the amount taken from 'len' in any given
iteration and changing the loop condition to reflect its unsigned nature.