Add solution to p11.
Create an 'encryption oracle' function per specifications in problem11.  One
which half the time encrypts using ecb and the other encrypts using cbc.  Also
create a test version of this function, called by it which deterministically
encrypts using one mode or the other.  Verify in tests both that it is
generating random (or at not identical) ouptuts and that our detection function
'guesses' the right one 100% of the time.