Move p11 oracle functions to matasano_problems and add p12 oracle.
The p11 oracle functions were written specifically to the p11 requirements and
likely are not going to be reusable in the context of block ciphers.  Move them
to the matasano_problems code to reflect their special-case-ness.  Also add
the new oracle for p12, which encouraged this move.

Add a test file for these functions to move the tests here out of BlockTest as
well.