constify BinaryBlob.
Too long have we been lax with our interfaces, but no more! Make sure we use
const to properly document and enforce BinaryBlob API which should not change
the underlying data.

Also, while testing this, we discovered a copy/paste error in one of the
constructors.  Remove it and add a B64 encoding test.